requirements.txt
pyyaml~=5.4.1
setuptools~=53.0.0
prometheus-client~=0.9.0
